Brings 42 commits of main into the 50-port branch. Six conflicts; every
resolution below was decided from the binary listing, not by preferring a side.

src/game/city_ui/TArmoryView.cpp, TShipyardView.cpp -- took this branch's bodies.
Main left 0x4ceda0 unclaimed (33% stub) with a comment asserting "no standalone
ctor address exists", and zeroed only productionView98. The listings refute that:
both 0x4ceda0 and 0x4c82c0 are real 32-byte out-of-line ctors (base ctor call,
vptr 0x652b10 / 0x651b30, one `xor eax,eax` reused for BOTH [this+0x94] and
[this+0x98], `mov eax,esi` return-this). Zeroing both fields and claiming the
addresses pairs each at 90.00%. Main's separate worry that 0x4c82c0 could not
pair uniquely is explained by the same defect: with only productionView98 stored
the body was byte-identical to its TBuildingView siblings'. Stale comments
replaced with the listing evidence.

config/original_entities.csv -- combined both sides rather than choosing. Main
pruned junk `global` label rows and relocated 6a3498 into the sorted globals
block; this branch consolidated interior fragment rows into their parent
functions (4cb8a0 301->506, 4cf350 287->492, 56ae10 416->810) and supplied real
prototypes. Kept both; verified no duplicate addresses and exactly one 6a3498 row.

tools/reccmp/progress_stats.py -- took main. This branch had added
`# pipe-split-ok` suppressions to the hand-rolled split; main deleted that code
path entirely in favour of tools.common.pipe_csv.read_pipe_rows, which achieves
the same end at the source. The superseded version is preserved in the preceding
commit so the intent is not lost from history.

docs/reference/{scalar-type-audit,turn-event-coverage}.md -- regenerated, not
hand-merged (generated artifacts).

0x004a2610 BuildArmyActionLabelFromLocalizationAndCounts -- both branches ported
this independently; duplicate claim removed by deleting this branch's standalone
TU and keeping main's TArmyMgr.cpp version, which is better typed (CStr255*),
sits beside its two callers, and ships with the 0x4a2900 caller. This branch's
copy scored higher in isolation but is wrong on the evidence: at 0x4a269b the
original loads the CString varargs with MOV (the m_pchData value, i.e. LPCSTR)
and uses LEA only for the CString* out-param, so main's static_cast<LPCSTR>(...)
is correct and this branch's &countStr passed the wrong thing. reccmp reports
this address `inconclusive` (unsupported control flow), which is not evidence of
a source defect, so the raw percentage was not chased.

src/game/military_ui/TBattleReportView.cpp -- auto-merged textually but broke the
build (C2679/C2664 x5): main retyped MapOrderBattleSideRecord's nameBuffer0c /
overlayLabel4c to CStr32[2] / CStr255[2], so indexing yields a struct, not char*.
Reached the buffer via .data, matching main's own usage elsewhere. No casts added.

Verification: build exit 0; gates exit 0; 554 tooling tests OK (4 skipped).
Stats improve on every metric -- exact 3792 -> 3809 (+17), not-exact 4220 -> 4182
(-38), exact/original +0.34 pp, size-weighted +1.28 pp, average +0.46 pp,
118 improved / 9 newly paired. The ~19 sub-1pp-to-few-pp dips are generated UI
builders in files untouched by this merge; all triage `inconclusive`, i.e. the
documented whole-image codegen wobble from combining two large branches, so the
stats baseline was refreshed rather than the source contorted.

+++ include/game/city_ui/TCountry.h
@@ -69,7 +69,7 @@ public:
   virtual void AddRegionIdToNationOwnedRegionList(int regionId);
   virtual void SetNationPercentFieldByModeAndDescriptorLinks(int targetNationSlot, int policyCode);
   virtual void DecrementDiplomacyCounterA2ByValue(int delta);
-  virtual int SumDiplomacyState1c6AndRelationDeltaSnapshot(short nationSlot);
+  virtual short SumDiplomacyState1c6AndRelationDeltaSnapshot(short nationSlot);
   virtual short GetDiplomacyCounterA2(void);
   virtual short GetDiplomacyExternalStateByTarget(short nationSlot);
   virtual short QueryNationMetricBySlot7C(short metricSlot);
